Move PortAvailability to network:io module.
