+++ infrastructure/io/src/main/java/tech/pegasys/teku/infrastructure/io/PortAvailability.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.util.config;
+package tech.pegasys.teku.infrastructure.io;
 
 import java.io.IOException;
 import java.net.DatagramSocket;
